Move rating code to dedicated field-rating view module
Create a proper field-rating view module following the pattern of other
field modules (field-phone, slider-field, etc.). The module:
- Exports helper functions (getRating, getRatingDisplayValue, isRatingValue)
- Registers the watchRatingIcons callback in the jetpack/form store
- Imports renderRatingIconsHtml from the rating-icons.js file

The main form view.js now imports helpers from the field-rating module
and no longer contains the watchRatingIcons callback.
